+++ src/main/java/restapi/webapp/services/UserService.java
@@ -1,6 +1,5 @@
 package restapi.webapp.services;
 
-import org.apache.catalina.User;
 import restapi.webapp.dtos.UserDTO;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
@@ -21,6 +20,7 @@ import restapi.webapp.repos.UserRepo;
 
 import java.util.HashMap;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.function.Function;
 import java.util.stream.Collectors;
@@ -79,7 +79,11 @@ public class UserService {
      */
     public ResponseEntity<?> getAllUsers(){
         List<UserEntity> userEntities = userRepo.findAll();
-        userEntities.stream().findAny().orElseThrow(() -> new UsersNotFoundException());
+        userEntities.stream().findAny().orElseThrow(UsersNotFoundException::new);
+
+        /* Explicitly declaring CollectionModel and not using getCorrespondingEntityType method because
+        getAllUsers should return a CollectionModel in any case.
+        */
         CollectionModel<EntityModel<UserEntity>> users = assembler.toCollectionModel(userEntities);
         return ResponseEntity.of(Optional.of(users));
     }
@@ -90,7 +94,10 @@ public class UserService {
      * @return ResponseEntity of corresponding message.
      */
     public ResponseEntity<?> deleteUserById(@NonNull Long id) {
-        userRepo.findById(id).stream().findAny().orElseThrow(() -> new UserNotFoundException(id));
+        UserEntity userEntity = userRepo.getUserEntityByUserId(id).get(0);
+        if(Objects.isNull(userEntity)){
+            throw new UserNotFoundException(id);
+        }
         userRepo.deleteUserEntityByUserId(id);
         return ResponseEntity.ok("User with the ID: " + id + " has been deleted.");
     }
@@ -101,7 +108,10 @@ public class UserService {
      * @return ResponseEntity of corresponding message.
      */
     public ResponseEntity<?> deleteUserByEmail(@NonNull String email) {
-        userRepo.getUserEntityByEmail(email).stream().findAny().orElseThrow(() -> new UserNotFoundException(email));
+        UserEntity userEntity = userRepo.getUserEntityByEmail(email).get(0);
+        if(Objects.isNull(userEntity)){
+            throw new UserNotFoundException(email);
+        }
         userRepo.deleteUserEntityByEmail(email);
         return ResponseEntity.ok("User with the email: " + email + " has been deleted.");
     }
@@ -113,7 +123,11 @@ public class UserService {
      * @return ResponseEntity of the created user
      */
     public ResponseEntity<?> createUser(@NonNull UserEntity user){
-        if (!userRepo.getUserEntityByEmail(user.getEmail()).isEmpty()) { throw new UserExistsException(user.getEmail()); }
+        // In case user already exists, don't override it.
+        UserEntity userEntity = userRepo.getUserEntityByEmail(user.getEmail()).get(0);
+        if (!Objects.isNull(userEntity)) {
+            throw new UserExistsException(user.getEmail());
+        }
 
         AvatarEntity currentAvatarEntity = user.getAvatarEntity();
         currentAvatarEntity.setSeed(user.getEmail());
@@ -133,7 +147,10 @@ public class UserService {
      */
     public ResponseEntity<?> updateUser(@NonNull UserEntity user) {
         // In case there's already a user with same credentials, it will save the changes.
-        userRepo.findById(user.getUserId()).stream().findAny().orElseThrow(() -> new UserNotFoundException(user.getUserId()));
+        UserEntity userEntity = userRepo.getUserEntityByUserId(user.getUserId()).get(0);
+        if (!Objects.isNull(userEntity)) {
+            throw new UserNotFoundException(user.getUserId());
+        }
         userRepo.save(user);
         log.info("User {} has been updated", user.getUserId());
         return ResponseEntity.of(Optional.of(assembler.toModel(user)));
@@ -150,7 +167,8 @@ public class UserService {
     public ResponseEntity<?> getUsersByLocation(@NonNull String city, @NonNull String streetName,
                                                 @NonNull String streetNumber, @NonNull String country){
         List<UserEntity> users = userRepo.getUserEntitiesByLocation(city, streetName, streetNumber, country);
-        users.stream().findAny().orElseThrow(() -> new UserNotFoundException(String.format("City %s, Street %s %s, Country %s", city, streetName, streetNumber, country)));
+        users.stream().findAny().orElseThrow(() -> new UsersNotFoundException
+                (String.format("City %s, Street %s %s, Country %s", city, streetName, streetNumber, country)));
         return getCorrespondingEntityType(users);
     }
 
@@ -175,7 +193,7 @@ public class UserService {
      */
     public ResponseEntity<?> getUserBySpecificParameter(@NonNull String param, @NonNull String value) {
         List<UserEntity> userEntities = this.methodsByParamsMap.get(param).apply(value);
-        userEntities.stream().findAny().orElseThrow(() -> new UserNotFoundException(String.format("%s %s", param, value)));
+        userEntities.stream().findAny().orElseThrow(() -> new UsersNotFoundException(param, value));
         return getCorrespondingEntityType(userEntities);
     }
 
@@ -187,10 +205,10 @@ public class UserService {
      * @param startingChar Starting character of user's last name to be checked.
      * @return ResponseEntity of the requested user, if exists.
      */
-    public ResponseEntity<?> getUserByAgeAndName(@NonNull Integer lower, @NonNull Integer upper, @NonNull String startingChar){
+    public ResponseEntity<?> getUsersByAgeAndName(@NonNull Integer lower, @NonNull Integer upper, @NonNull String startingChar){
         List<UserEntity> userEntities = this.userRepo.getUserEntityByAgeBetweenAndLastNameStartingWith(lower,
                 upper, startingChar);
-        userEntities.stream().findAny().orElseThrow(() -> new UsersNotFoundException());
+        userEntities.stream().findAny().orElseThrow(UsersNotFoundException::new);
         return getCorrespondingEntityType(userEntities);
     }
 
@@ -201,6 +219,12 @@ public class UserService {
      * @return ResponseEntity of the requested user, if exists.
      */
     public ResponseEntity<?> getUserDtoInfo(@NonNull Long id) {
+        /*
+        This method uses the pre-defined repo method findById because it returns an Optional<T>,
+        which allows us to map the returned value into a UserDTO and return an exception easily,
+        in case of an error.
+         */
+
         return userRepo.findById(id)
                 .map(UserDTO::new)
                 .map(userDTOAssembler::toModel)
@@ -215,7 +239,8 @@ public class UserService {
      */
     public ResponseEntity<?> getAllUsersDtoInfo() {
         List<UserEntity> userEntities = userRepo.findAll();
-        userEntities.stream().findAny().orElseThrow(() -> new UsersNotFoundException());
+        // Check if there are any users that exist in DB
+        userEntities.stream().findAny().orElseThrow(UsersNotFoundException::new);
         return ResponseEntity.ok(userDTOAssembler.toCollectionModel(
                         userEntities
                         .stream()
